refactor, move findScopePath from legacy to a new component
