+++ scopes/harmony/config/config.main.runtime.ts
@@ -11,7 +11,7 @@ import LegacyWorkspaceConfig, {
   WorkspaceConfigLoadFunction,
 } from '@teambit/legacy/dist/consumer/config/workspace-config';
 import { PathOsBased, PathOsBasedAbsolute } from '@teambit/legacy/dist/utils/path';
-import { findScopePath } from '@teambit/legacy/dist/utils';
+import { findScopePath } from '@teambit/scope.modules.find-scope-path';
 import { MainRuntime } from '@teambit/cli';
 import { GlobalConfig, Harmony } from '@teambit/harmony';
 import path from 'path';
