+++ src/utils/index.ts
@@ -6,7 +6,6 @@ import { checksum, checksumFile } from './checksum';
 import getWithoutExt from './fs/fs-no-ext';
 import getExt from './fs/get-ext';
 import isDirEmpty from './fs/is-dir-empty';
-import { pathHas, pathHasAll, findScopePath } from './fs/propogate-until';
 import readDirIgnoreDsStore, { readDirSyncIgnoreDsStore } from './fs/read-dir-ignore-ds-store';
 import glob from './glob';
 import retrieveIgnoreList from './ignore/ignore';
@@ -67,9 +66,6 @@ export {
   isBitUrl,
   isDir,
   resolveHomePath,
-  findScopePath,
-  pathHas,
-  pathHasAll,
   isDirEmpty,
   isDirEmptySync,
   immutableUnshift,
